% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_get.R
\name{archive_get}
\alias{archive_get}
\title{Download batches of GRIB files from the RAP archive or GFS}
\usage{
archive_get(
  model,
  date_rel,
  hour_rel = seq(0, 23, by = 2),
  hour_pred = 1L,
  aoi = NULL,
  grib_dir = NULL,
  quiet = FALSE,
  overwrite = FALSE,
  alternate = TRUE
)
}
\arguments{
\item{model}{character, either 'rap_archive' or 'gfs_0p25'}

\item{date_rel}{Date vector, the forecast release date(s)}

\item{hour_rel}{integer vector, the forecast release hour}

\item{hour_pred}{integer vector, the prediction hours}

\item{aoi}{geometry object defining the area of interest, passed to \code{archive_url}}

\item{grib_dir}{character, path to storage directory for GRIB/GRIB2 files}

\item{quiet}{logical, prints progress information to console}

\item{overwrite}{logical, whether to overwrite existing files found in \code{grib_dir}}

\item{alternate}{logical, if TRUE and a file is not found, an alternate is attempted}
}
\value{
data frame with info about the requested times, and results of download attempts
}
\description{
This is a loosely based on \code{rNOMADS::ArchiveGribGrab}, but customized for our use case.
It downloads GRIB/GRIB2 file(s) from NCEI or NCEP, from a particular release date-time
(\code{date_rel}, \code{hour_rel}) and prediction hour (\code{hour_pred}), or a sequence of them.
}
\details{
The available choices for \code{model} are:
\itemize{
\item 'rap_archive' : the archived RAP/RUC files at NCEI
\item 'gfs_0p25' : the most current GFS release and previous few days
}

Each of \code{date_rel}, \code{hour_rel}, and \code{hour_pred} can be vectors and all times are expected
in the UTC time zone. The function matches requests to files based on these three arguments,
and \code{overwrite=FALSE} will skip requests where a matching forecast file was found already
on disk.

With the 'rap_archive' model there can be multiple matches for a given request: grb2
files are always preferred with .grb (version 1) only used as a fall-back. Similarly
rap_130 files are preferred over the (older model and coarse resolution) ruc_252 files.

Note that with 'rap_archive' the only valid choices for \code{hour_pred} are 0 and 1, but
\code{hour_rel} can be anything in \code{seq(24)}. With 'gfs_0p25' the only valid choices for
\code{hour_rel} are 0, 6, 18, but \code{hour_pred} can be anything in \code{seq(120)}. To get a list
of all choices available on a given date, see \code{?archive_url}

Requests for multiple files are processed in a loops that should continue in spite of
download errors. If an error occurs in the middle of a download, the function will delete
the (probably corrupt) file and continue on to the next. Multiple requests to NCEI/NCEP
servers are spaced out using \code{Sys.sleep} such that the request rate never exceeds
60 / minute.

If \code{alternate=TRUE} and \code{hour_pred=1}, the function will attempt to download the
0-hour ahead forecast whenever the 1-hour file is missing from the NCEI server. Note
that this has only been implemented for the default \code{hour_rel} and \code{hour_pred}
}
