% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_pad.R
\name{archive_pad}
\alias{archive_pad}
\title{Expands/subsets the tibble returned by \code{grib_list} to get gap-less time series}
\usage{
archive_pad(grib_df, times = "posix_pred", until = NULL, quiet = FALSE)
}
\arguments{
\item{grib_df}{data frame containing column \code{t}}

\item{times}{character column name containing \code{POSIXct} times.}

\item{until}{POSIXct time, the series is truncated/padded to this time point}

\item{quiet}{logical indicating to suppress console output}
}
\value{
a tibble with the same columns as \code{grib_df}
}
\description{
This creates a time series with interval \code{step_hours}, starting from the earliest time
(\code{times} column) found in \code{grib_df}. Rows inconsistent with the series are removed, and
new rows (with NA fields apart from \code{times}) are added to for missing times.
}
