% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_update.R
\name{archive_update}
\alias{archive_update}
\title{Initialize or update a local archive of forecasts from RAP/RUC or GFS}
\usage{
archive_update(
  base_dir,
  from = NULL,
  to = NULL,
  hour_rel = 0L,
  hour_pred = 1L,
  model = "rap_archive",
  aoi = NULL
)
}
\arguments{
\item{base_dir}{path for the GRIB storage subfolder}

\item{from}{Date, the first date in the sequence}

\item{to}{Date, the last date in the sequence}

\item{hour_rel}{integer vector, the release hours}

\item{hour_pred}{integer vector, the prediction hours}

\item{model}{character, either 'rap_archive' or 'gfs_0p25'}
}
\value{
a data frame of information about the files downloaded
}
\description{
This creates the subfolder 'gribs' in your \code{base_dir} if it doesn't exist
already, and fills it with archived forecast GRIB files from the selected
model.
}
\details{
Two models are currently supported: 'rap_archive' downloads the Rapid Refresh
(RAP) and Rapid Update Cycle (RUC, the predecessor of RAP) from NCEI;
and 'gfs_0p25' fetches the Global Forecast System (GFS) from NCEP. Available
release times for the RAP/RUC range from from 2005-01-01 until two days before
present; and for GFS there is a rolling window of availability from about 10
days before present until the present day.

To download all available times from GFS, leave the \code{from}, \code{to}, and \code{hour_rel}
arguments \code{NULL}. Otherwise they specify the start and end dates, and the release
hours to look for.

\code{from} and \code{to} can be NULL in RAP/RUC calls, in which case defaults are assigned
based on the date range of existing files in your archive. If there are no existing
files, \code{from} is set to the earliest available date (\code{.from_def}), otherwise it is
set to latest date among the existing files. The default for \code{to} is latest
available date.

The time at which a forecast becomes/became valid is equal to the release time
\code{hour_rel} plus the prediction hour \code{hour_pred}. These can be vectors if you want
to select sequences on every date. See \code{?archive_get} (which does all the work)
for details on availability, and how multiple versions of forecasts are handled.
}
