% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_url.R
\name{archive_url}
\alias{archive_url}
\title{Get a list of download URLs for GRIB files at NCEI and NCEP}
\usage{
archive_url(date_rel, model = "rap_archive", aoi = NULL, quiet = FALSE)
}
\arguments{
\item{date_rel}{Date, the forecast release date of interest}

\item{model}{either 'rap_archive' or 'gfs_0p25'}

\item{aoi}{a geometry object, the area of interest (for cropping 'gfs_0p25')}

\item{quiet}{logical suppresses console output}
}
\value{
a tibble containing URLs and information about the files behind them
}
\description{
This fetches download links for files from the RAP/RUC archive at NCEI,
and the (current) GFS forecasts from NCEP. The date arguments specify the
release date (when the file was published) and the function returns links for
all forecast hours at all available release times.
}
\details{
IMPORTANT NOTE: be polite and use the function in moderation! See below:

Each function call causes \code{readLines} to request the HTML content of a webpage from
NCEP or NCEI. Users who make requests too frequently can have their IP addresses
banned by these servers! There is a 120 requests/minute limit, and the guidance from
NCEP is to aim for around 50 requests/minute when making requests in a loop. Enforce
this in your own code using \code{proc.time()} and \code{Sys.sleep()} as needed.

When \code{model='gfs_0p25'}, the argument \code{aoi} is used to find appropriate bounding
box, and the resulting download links should produce a suitably cropped raster.
Only the first 120 prediction times (hours) are returned. GFS forecasts are released
4x daily, so you can expect 1-4 links for each of the 120 prediction hours
depending on what time of day you make the call.

\code{aoi} is ignored when \code{model='rap_archive'}, which doesn't support cropping.
All of its files cover all of North America. With this dataset, only the 0
and 1-hour prediction times are available, and releases happen hourly, so you can
typically expect 2 links for each of 24 release hours.

with both datasets, the links are for files containing all variables at all levels.
GFS supports requests for a subset of variables, so maybe we can add functionality
for that later on. All times are in UTC.
}
