% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_downscale.R
\name{.nc_downscale}
\alias{.nc_downscale}
\title{Down-scale a spatio-temporal time series}
\usage{
.nc_downscale(g, z_obs, pars, X, poly_out = NULL)
}
\arguments{
\item{g}{sk grid containing the locations of observed grid points (see \code{?snapKrig::sk_cmean})}

\item{z_obs}{list of vectors, the observed data at different times}

\item{pars}{list of spatial model parameters (eg as returned by \code{snapKrig::sk_fit})}

\item{X}{numeric matrix of covariates passed (see \code{?snapKrig::sk_cmean})}

\item{poly_out}{sf POLYGON objects to use for cropping output (or NULL for no cropping)}
}
\value{
either a SpatVector, a list of them
}
\description{
This uses \code{snapKrig} and a previously fitted model (parameters list \code{pars}) to interpolate
from a grid of observed data onto a grid of finer resolution.
}
\details{
The target grid and the location of the observations within it are specified by \code{g}, in
the same way as they are specified in \code{snapKrig::sk_cmean}; NAs for unobserved grid points
and non-NA values otherwise. The values themselves are ignored, and the data from \code{z_obs}
are substituted in their place.

\code{z_obs} should be a list of vectors, one for each time layer to process. These vectors
should contain only the observed (non-NA) data, in the vectorized ordering used by \code{snapKrig}.
For example to downscale the data layer in \code{g}, you could set \code{z_obs = list(g[!is.na(g)])}.

The data matrix \code{X} should be ordered the same way, but it must contain a row for every point
in the target grid, including unobserved ones. Typical usage is to create \code{X} with a call to
\code{space_X} (with \code{intercept=FALSE}).

Output layers are returned in a SpatRaster. If \code{poly_out} is supplied, the function crops
all layers to the bounding box of the polygons as a whole.
}
