% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_update.R
\name{dummy_nc}
\alias{dummy_nc}
\title{Make an empty NetCDF file covering time period prior to a cutoff date}
\usage{
dummy_nc(output_nc, grib_df, aoi, date_cutoff = as.Date("2016-10-01"))
}
\arguments{
\item{output_nc}{the output file name}

\item{grib_df}{data frame returned by \code{grib_list(..., dupe=FALSE)}}

\item{aoi}{geometry object passed to \code{grib_idx} (area of interest)}
}
\value{
nothing, but overwrites the file \code{output_nc}
}
\description{
The function writes a NetCDF file containing an empty layer for each time in
\code{grib_df[['posix_pred']]} prior to \code{date_cutoff}. All data values are set to \code{NA}.
}
\details{
WARNING: This overwrites any existing data in \code{output_nc}!

\code{pcp_total} is not currently available prior to 2016-10-01 (the default for
\code{date_cutoff}). The purpose of this function is to trick \code{nc_update} into not
checking for this variable on prior dates. This helps speed up the initial call
to \code{nc_update}.

At least one of the files listed in \code{grib_df} must exist so that this function can
get a template grid from it. It is assumed that all files in \code{grib_df} are from
same spatial grid. There are three different grids: GFS, and RAP/RUC coarse
and fine - make sure you filter to only of these groups.
}
