% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_extract.R
\name{estimate_memory}
\alias{estimate_memory}
\title{Estimate memory usage in GB for a nested list of numeric matrices}
\usage{
estimate_memory(gdim, n_var = 1L, n_file = 1L, quiet = FALSE)
}
\arguments{
\item{gdim}{integer vector, the grid dimensions (order unimportant)}

\item{n_var}{integer number of variables}

\item{n_file}{integer number of files}

\item{quiet}{logical, if FALSE the function prints information to console}
}
\value{
numeric (GB)
}
\description{
Helper function for \code{grib_extract}
}
\details{
This prints a rough estimate of the memory needed for a list of \code{n_file * n_var}
grids of dimensions \code{gdim}. The output is always in GB whereas the console message
is printed in MB if the volume is less than 1 GB.
}
