% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_extract.R
\name{grib_idx}
\alias{grib_idx}
\title{Find layer and sub-grid indices for calls to \code{terra::rast} and \code{terra::values}}
\usage{
grib_idx(grib_path, regex = NULL, aoi = NULL, quiet = FALSE, try_again = FALSE)
}
\arguments{
\item{grib_path}{character (or dataframe with 'path' column). The path to the GRIB file}

\item{regex}{character vector, regular expressions to match to layer names (using \code{grepl})}

\item{aoi}{geometry object coercible to \code{SpatVector}, the area of interest}

\item{quiet}{logical indicating to suppress console output}

\item{try_again}{logical indicating to attempt to load another file if the first fails}
}
\value{
list of indices (see details)
}
\description{
This uses \code{terra} to read the layer names and grid dimensions from the GRIB at
\code{grib_path} (but not its data values) in order to learn (1) the index of layer
names matching the regular expressions in \code{regex}, and (2) the dimensions and
position of the sub-grid inscribing the area of interest \code{aoi}. These vectors can
then be passed to \code{terra::rast} and \code{terra::values} to speed up loading the data
(see \code{grib_read}).
}
\details{
The function returns a list with the following
\itemize{
\item 'k' = index of requested layers
\item 'ij' = row/column of top-left corner point in AOI sub-grid
\item 'dim' = dimensions of AOI sub-grid
\item 'r_aoi' = empty SpatRaster defining the AOI
}

If \code{regex=NULL}, all layers are returned. If any of the \code{regex} match more than
one layer, only the first match is returned. If \code{aoi=NULL}, the entire grid is
requested. Otherwise \code{aoi} should be coercible to 'SpatVector' and must overlap
with some part of the grid in \code{grib_path} (after transforming to a common CRS).

\code{grib_path} can be a data-frame containing multiple paths, in which case the first
is loaded. If this load attempt fails, and \code{try_again=TRUE}, the function will then
attempt to load the second file, and so on.
}
