% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_list.R
\name{grib_list}
\alias{grib_list}
\title{Create/update a CSV with info on GRIB files retrieved with \code{archive_get}.}
\usage{
grib_list(grib_dir, csv = NULL, dupe = NA, quiet = FALSE)
}
\arguments{
\item{grib_dir}{path to directory with GRIB files}

\item{csv}{character, file name for the CSV (set \code{NULL} to use default in \code{wx_file})}

\item{dupe}{logical, indicating to keep only the first of every duplicate time}

\item{quiet}{logical, suppresses console output}
}
\value{
a tibble, the contents of the CSV
}
\description{
The CSV is saved to \code{file.path(base_dir, csv)}, overwriting anything already there,
except when \code{csv=NA}, in which case the function returns its usual output but writes
nothing to disk.
}
\details{
The data frame reports information about the GRIB files in \code{grib_dir} parsed from their
file names. Parsing is slow with tens of thousands of files, so the results are saved in
a CSV on disk for quick access later on. Each time the function is called it checks the
directory for changes, updating the CSV on disk as needed (unless \code{csv=NA}).

Output fields describe the file, the model it came from, and the grid resolution.
Three fields describe the (two) temporal dimensions of the data:
\itemize{
\item posix_pred : the POSIXct time at which the forecast file is valid
\item date_rel : the Date on which the forecast was created
\item hour_rel : the (integer) hour of the day when the forecast was created
}

Output rows are first sorted by \code{posix_pred} (chronological order), then by \code{date_rel},
then \code{hour_rel}. Duplicate \code{posix_pred} values are dealt with according to the value of
\code{dupe}. The default \code{NA} returns all duplicates; \code{FALSE} returns only the first of every
duplicate (ie the most recently released file); and \code{TRUE} returns the complement of
\code{FALSE} (ie all duplicates except the most recent).

Call \code{wx_file('csv')} to get the default \code{csv}; and \code{wx_file('grib', base_dir)}
to get the default \code{grib_dir} for project directory \code{base_dir}.
}
