% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{my_archive_padder}
\alias{my_archive_padder}
\title{Expands/subsets the tibble returned by \code{my_archive_lister} to get gap-less time series}
\usage{
my_archive_padder(grib_df, times = "posix_pred", until = NULL, quiet = FALSE)
}
\arguments{
\item{grib_df}{data frame containing column \code{t}}

\item{times}{character column name containing \code{POSIXct} times.}

\item{quiet}{logical indicating to suppress console output}
}
\value{
a tibble with the same columns as \code{grib_df}
}
\description{
This creates a time series with interval \code{step_hours}, starting from the earliest time
(\code{times} column) found in \code{grib_df}. Rows inconsistent with the series are removed, and
new rows (with NA fields apart from \code{times}) are added to for missing times.
}
