% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{my_detect_step}
\alias{my_detect_step}
\title{Find the step size in a time series as minimal time difference between observations}
\usage{
my_detect_step(grib_df, times = "posix_pred", quiet = FALSE, more = FALSE)
}
\description{
Returns the step size in hours, or if \code{more=TRUE}, appends the columns "ts_hour" (the
number of hours since the first observation) and "interval" (the difference in
"ts_hour" between a given observation and the previous). The step size is the least
nonzero "interval".
}
