% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{my_dummy_nc}
\alias{my_dummy_nc}
\title{Make an empty NetCDF file covering time period prior to a cutoff date}
\usage{
my_dummy_nc(output_nc, grib_df, aoi, date_cutoff = as.Date("2016-10-01"))
}
\arguments{
\item{output_nc}{the output file name}

\item{grib_df}{data frame returned by \code{grib_list(..., dupe=FALSE)}}

\item{aoi}{geometry object passed to \code{grib_idx} (area of interest)}
}
\value{
nothing, but overwrites the file \code{output_nc}
}
\description{
Note that this overwrites any existing data in \code{output_nc}. The result is a file
containing a layer for each time in \code{grib_df[['posix_pred']]} prior to \code{date_cutoff},
with all data values set to \code{NA}.
}
\details{
At least one of the files listed in \code{grib_df} must exist so that this function can
read the grid dimensions etc. It is assumed that all files in \code{grib_df} are of the
same spatial grid.

\code{pcp_total} is not available prior to 2016-10-01 (the default for \code{date_cutoff}).
This function allows us to trick \code{nc_update} into not checking for this variable
on prior dates which speeds loading.
}
