% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{my_fit_spatial}
\alias{my_fit_spatial}
\title{Fit a spatial covariance model to a NetCDF time series}
\usage{
my_fit_spatial(
  var_nm,
  base_dir,
  dem_path,
  input_nm = "fine",
  n_max = 500,
  append = TRUE,
  positive = NULL
)
}
\arguments{
\item{var_nm}{character vector or list, the names of the variables to fit (in a loop)}

\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{dem_path}{character path to the DEM (passed to \code{terra})}

\item{input_nm}{character vector or list, subdirectories containing the .nc files to fit}

\item{n_max}{integer, maximum number of layers to sample for fitting}

\item{append}{logical indicating to keep old model fits in the JSON}

\item{positive}{logical indicating to exclude all-zero layers}
}
\value{
returns nothing, but writes to JSON files in \code{file.path(base_dir, train_nm)}
}
\description{
This uses \code{snapKrig} to fit a spatial MVN covariance model to a random
subsample of the layers in a NetCDF file (up to \code{n_max}). The covariates matrix
is generated by passing the DEM at \code{dem_path }to \code{my_space_X}. It includes a
spline basis for lapse rate, and columns for geographic coordinates.
}
\details{
This uses a purely spatial geo-statistical model with a nugget effect, and a
stationary Gaussian covariogram with geometric anisotropy (oriented along the
cardinal axes). Time layers are treated as mutually independent, and the function
uses a random subsample (<=\code{n_max}) of times for model fitting.

Subsampling excludes layers with any number of NAs. when \code{positive=TRUE}, it also
excludes all-zero layers (useful for precipitation). When \code{positive=NULL} (the
default), the function detects precipitation layers automatically (and sets
\code{positive=TRUE}) by checking if any of the variable names begin with 'pcp'.

\code{input_nc} can be a vector of sub-directories, and \code{var_nm} can be a list of
character vectors (specifying sets of equivalent names), allowing users to specify
multiple source files for a single variable, similar to \code{nc_resample}.

Variable names (\code{var_nm}) specify a subset of the nc files in \code{input_nm} to process
in a loop. The results are written to a JSON file in sub-directory "model" of the first
sub-directory listed in \code{input_nm}. Get the list of all output paths with:

\verb{file_wx('spatial', base_dir, input_nm[1], as.list(names(var_nm))}

If \code{append=FALSE} the function deletes any existing data in the JSON and replaces
it with the new model fit. If \code{append=TRUE}, the new fit is appended to the bottom of
the existing list in the file. The results include the output of \code{my_sk_fit}, along
with the time of the function call, and the input files and variable names.
}
