% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_helpers.R
\name{my_fit_temporal}
\alias{my_fit_temporal}
\title{Fit a temporal model to a NetCDF time series}
\usage{
my_fit_temporal(
  var_nm,
  base_dir,
  input_nm = "fine",
  daily_n = 5L,
  yearly_n = 5L,
  knots_n = 5L,
  p_max = 0.1,
  n_max = NA
)
}
\arguments{
\item{var_nm}{list of character vectors, the variable names to fit}

\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{input_nm}{character vector or list, subdirectories containing the .nc files to fit}

\item{p_max}{numeric the maximum proportion of missing points in the series}

\item{n_max}{the maximum number of points to sample (NA to select all)}
}
\description{
Fit a temporal model to a NetCDF time series
}
