% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{my_gap_finder}
\alias{my_gap_finder}
\title{Return a tibble of start/end indices for all NA sequences in a vector}
\usage{
my_gap_finder(x, start_only = FALSE, times = NULL, invert = FALSE)
}
\arguments{
\item{invert}{}
}
\value{
a tibble containing a row for each distinct gap
}
\description{
Identifies the start and end points of all gaps (contiguous stretches of \code{NA} values)
in a vector, along with their lengths. If \code{invert=TRUE} the function returns
information about the non-\code{NA} segments.
}
\details{
\code{x} can be any vector containing \code{NA} entries. If \code{times} is supplied, The function
returns the value of \code{times} at the start/end index, instead of the index itself.

If a data frame is passed to \code{x} then \code{times} should be a (character) column
name identifying the time values to use - the first (non-time) column in \code{x}
is then checked for \code{NA}s.

If there are no \code{NA}s (or, with \code{invert=TRUE}, no non-NA's), the function returns NULL
}
