% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_read_helpers.R
\name{my_grib_read}
\alias{my_grib_read}
\title{GRIB loader to catch errors with missing/damaged files}
\usage{
my_grib_read(grib_path, regex = NULL, aoi = NULL, quiet = FALSE)
}
\arguments{
\item{grib_path}{character path to the GRIB file to (attempt to) load}

\item{regex}{character vector passed to \code{my_grib_idx} (layer names)}

\item{aoi}{geometry object passed to \code{my_grib_idx} (area of interest)}

\item{quiet}{logical indicating to suppress console output}
}
\value{
SpatRaster containing the requested layers and sub-grid
}
\description{
This is a wrapper for \code{terra::rast} that loads only the sub-grid and
layers specified in \code{regex} and \code{aoi} (using \code{my_grib_idx}) for the GRIB
file at \code{grib_path}.
}
\details{
This catches loading errors - instead of halting, the function
returns the error object (and prints it as a warning if \code{quiet=FALSE}).
}
