% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_helpers.R
\name{my_impute_temporal}
\alias{my_impute_temporal}
\title{Fill temporal gaps using AR(2) model and seasonal predictor}
\usage{
my_impute_temporal(
  var_nm,
  base_dir,
  input_nm = "fine",
  output_nm = "completed",
  model_nm = input_nm[1],
  n_max = NULL,
  until = NULL,
  quiet = FALSE
)
}
\arguments{
\item{var_nm}{character vector or list of them, the input variables to process}

\item{quiet}{}
}
\description{
Fill temporal gaps using AR(2) model and seasonal predictor
}
