% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_helpers.R
\name{my_nc_write}
\alias{my_nc_write}
\title{Modify a NetCDF file by adding new time layers}
\usage{
my_nc_write(r, p, overwrite = FALSE, append = TRUE)
}
\arguments{
\item{r}{SpatRaster with POSIXct vector \code{terra::time(r)}, the data to write}

\item{p}{character path the (.nc) time series data file to write}

\item{overwrite}{logical indicating to write changes to the nc file on disk}

\item{append}{logical indicating to only add new times to the file}
}
\value{
vector of POSIXct times, the layers added to the file
}
\description{
This creates or updates an existing NetCDF file by adding the layers (times)
in the supplied SpatRaster \code{r}.
}
\details{
The function does little checking to see if the request is valid. It assumes
that \code{p} is a valid path to a NetCDF file (.nc), and that raster grid in \code{r}
is compatible with the file at \code{p} in terms of dimensions etc.

\code{r} must have a time for each layer (check \code{terra::time(r)}). Only new times
(those not found in the file at \code{p}) are copied when \code{append=TRUE}. Set
\code{append=FALSE} to copy all times from \code{r}, replacing any existing ones. This
retains any times in the nc file that are not found in \code{r}.

If \code{overwrite=FALSE} (the default), the function overwrites nothing but
returns the times that would be added to the file.

When overwriting an existing file, the function initially writes to a temporary
file in the same directory. At the end of the function call the old file is
deleted and the temporary one is renamed to take its place. If something goes
wrong and the function halts, this temporary file can be safely deleted.

After writing to the .nc file, the function creates/updates the JSON attributes
file in the 'time' subdirectory.
}
