% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_workflow.R
\name{my_resample}
\alias{my_resample}
\title{Create/update a NetCDF archive of resampled coarse resolution grids}
\usage{
my_resample(
  var_nm,
  base_dir,
  input_nm = list(coarse = "coarse", fine = "fine"),
  output_nm = "coarse_resampled",
  r_fine = NULL,
  from = NULL,
  append = TRUE,
  ...
)
}
\arguments{
\item{var_nm}{character vector or list of them, names of the variables to process}

\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{input_nm}{list of character vectors, namess of the source file subdirectories in \code{base_dir}}

\item{output_nm}{character name for the sub-directory to create.}

\item{...}{additional arguments passed to \code{terra::project}}

\item{SpatRaster}{optional target grid for resampling}
}
\value{
nothing, but possibly modifies the nc and JSON files in \code{file.path(base_dir, output_nm)}
}
\description{
This resamples the coarse resolution time series to get data at fine resolution,
writing results to nc files in subdirectory \code{output_nm}. Only those times found in
the coarse series but not the fine series are included.
}
\details{
Specify the coarse and fine sub-directories as named elements in list \code{input_nm}. Each
should be a character vector of sub-directories to check for data on \code{var_nm}. When
multiple sub-directories contain data for a given time, the function extracts the data
from the first (according to the order in \code{input_nm})

\code{var_nm} can be a list of character vectors, where each vector gives a set of variable
names to be treated as equivalent (eg 'pcp' and 'pcp_total', which are the same variable,
compiled from different sources). The output variable is assigned the corresponding name
from \code{var_nm}. If \code{var_nm} is unnamed, the function names it using the first element of
each vector.

The sub-directory \code{output_nm} will be created if it doesn't exist. If the file in
\code{output_nm} exists already, only those times available in \code{input_nm}, but not already
present in \code{output_nm}, are appended.

Coarse resolution grids are by default resampled by bilinear averaging (with GDAL).
Change this by setting named arguments in \code{...} to pass to \code{terra::project}.
}
