% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{my_sk_fit}
\alias{my_sk_fit}
\title{Fit a spatial model and return a list of results for use in \code{my_fit_spatial}}
\usage{
my_sk_fit(p, X, n_max = 100, t_fit = NULL, positive = NULL)
}
\arguments{
\item{p}{character vector of input file paths to time series .nc files}

\item{X}{numeric matrix of spatial covariates (output of \code{my_space_X})}

\item{n_max}{integer > 0 number of layers to sample}

\item{t_fit}{character vector of POSIXct times to sample}

\item{positive}{logical indicating to exclude all-zero layers}
}
\value{
list of model fitting results
}
\description{
This is a helper function for \code{my_fit_spatial}. It fits a random subsample of
the observed data (from \code{p}) using the covariates in \code{X}. Generate \code{X} by passing
an example grid from \code{p} (along with a DEM) to \code{my_space_X}.
}
\details{
The function returns a list with the following
\itemize{
\item at : the UTC time of the function call
\item file : the files containing the sampled input data
\item pars : fitted parameters in a snapKrig-friendly list
\item bds : parameter bounds (lower, initial, fitted, upper)
\item betas : GLS estimates of the linear model coefficients
\item train : the times in the training set
}
}
