% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_export.R
\name{nc_aggregate}
\alias{nc_aggregate}
\title{Aggregate sub-daily data to daily}
\usage{
nc_aggregate(p, fun = "mean", tz = "UTC", origin_hour = 0L)
}
\arguments{
\item{p}{character vector path to the nc file(s)}

\item{fun}{function, one of "mean", "min", or "max"}

\item{tz}{character time zone for \code{origin_hour}}

\item{origin_hour}{integer, steps are aligned to start at this hour of the day}
}
\value{
SpatRaster, the aggregated data
}
\description{
Aggregate sub-daily data to daily
}
