% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_aggregate_space.R
\name{nc_aggregate_space}
\alias{nc_aggregate_space}
\title{Aggregate a gridded spatio-temporal dataset over user-specified regions}
\usage{
nc_aggregate_space(
  base_dir,
  input_nm = .nm_down,
  output_nm = .nm_export,
  var_nm = .var_daily,
  poly_in = NULL,
  fun = NULL,
  from = NULL,
  to = NULL,
  file_ext = ".tif"
)
}
\arguments{
\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{input_nm}{character, the sub-directory name of input to process}

\item{output_nm}{character, the sub-directory name for output}

\item{var_nm}{character vector, the variable names to process}

\item{poly_in}{polygon(s) to aggregate over, any object acceptable by \code{sf::st_geometry}}

\item{fun}{character, name of the function to apply for aggregation}

\item{from}{Date, the first date of the sequence to process}

\item{to}{Date, the last date of the sequence to process}

\item{file_ext}{character, the input file extension to load, either '.tif' or '.nc'}
}
\value{
returns nothing but writes CSV files to \code{output_nm}
}
\description{
This takes a set of input files (NetCDF or GeoTIFF) and applies the aggregation
function \code{fun} to its values at each time-step. If \code{poly_in} is supplied, the
aggregation happens separately over polygons.
}
\details{
The function writes its output to CSV files in the output directory \code{output_nm},
with each variable saved to its own sub-directory. A separate file is written for
each year, and results for different polygons are written different columns of the
output.

For consistency with earlier datasets in the workflow, this function also
writes JSON files indexing the dates in each CSV file (in subdirectories named "time").
Note also that sub-directories of \code{output_nm} will use the suffix '.nc' (even though
we aren't dealing with NetCDF at this point)

By default \code{from} is set to to 10 days before latest existing date in the output,
and \code{to} is set to latest available date in \code{input_nm}. If any part of a year is
included in this date range, then the whole year is processed (and the output file
is overwritten completely)

Non-NULL \code{fun} is passed directly to \code{terra::extract} to specify the aggregation
function. Otherwise the function looks for variable names given in snake case (with
'_' separating parts), and assigns the last part to \code{fun}. For example if \code{fun=NULL}
then variables named 'tmp_max' or 'tmp_mean_max' would both prompt \code{max} for
aggregation. If a variable name cannot be split in this way, the function reverts to
the default \code{fun='mean'}.
}
