% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_downscale.R
\name{nc_downscale}
\alias{nc_downscale}
\title{Down-scale a spatio-temporal time series in a set of NetCDF files}
\usage{
nc_downscale(
  base_dir,
  dem,
  down = 100,
  input_nm = .nm_daily,
  model_nm = .nm_model,
  output_nm = .nm_export,
  var_nm = .var_daily,
  poly_out = NULL,
  edge_buffer = NULL,
  fun = NULL,
  dates = NULL,
  write_nc = TRUE
)
}
\arguments{
\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{dem}{SpatRaster of elevation data at finer resolution than the input data}

\item{down}{integer, the down-scaling factor}

\item{input_nm}{character, the sub-directory name of input to process}

\item{model_nm}{character, the sub-directory name where the model files can be found}

\item{output_nm}{character, the sub-directory name for output}

\item{var_nm}{character vector, the name(s) of the variable to process}

\item{poly_out}{sfc object, polygon of interest}

\item{edge_buffer}{numeric, length in metres to buffer input data (see details)}

\item{dates}{Date vector, the dates to process (NULL for all)}
}
\value{
returns nothing but possibly modifes the NetCDF data in \code{output_nm}
}
\description{
This is a wrapper for \code{.nc_downscale} that down-scales - ie predicts at finer
resolution - the gridded time series data in \code{input_nm}. By default it writes
a set of (yearly) NetCDF files in the sub-directory \code{output_nm} of \code{base_dir},
with one set for each of the variables named in \code{var_nm}.
}
\details{
Down-scaling happens by universal kriging using the fitted spatial covariance
parameters saved in the folder named in \code{model_nm} (generate this using
\code{space_fit}), and using covariates generated from the digital elevation model
in \code{dem} (with values in metres).

Argument \code{down} is the (integer) factor by which the grid spacing is decreased:
eg. \code{down=2} "doubles the resolution" by inserting a new grid point in between
each existing pair of (adjacent) grid points.

Specify an area of interest with \code{poly_out}. The function conditions its
predictions on a subset of the input grid covering \code{poly_out}, buffered by
the distance \code{edge_buffer} (in m). By default this is set to the diagonal length
of a single grid cell. Setting this to a large number (eg the side length of your
domain) will ensure that all observations are considered in predictions.
}
