% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_export.R
\name{nc_export}
\alias{nc_export}
\title{Downscale and export completed time series by aggregating over areas}
\usage{
nc_export(
  base_dir,
  var_nm = NULL,
  output_nm = .nm_export,
  write_csv = FALSE,
  fun = "mean"
)
}
\arguments{
\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{var_nm}{character vector, the variable(s) to export (NULL to export all)}

\item{output_nm}{character, the sub-directory name for output}

\item{write_csv}{logical, Writes a CSV copy of the data, and point locations to geojson}

\item{fun}{character, a function name like "mean", "min", or "max"}
}
\value{
vector of file paths written to the \code{output_nm} directory
}
\description{
This exports all data each of the variables named in \code{var_nm} to their own (single)
NetCDF file in \code{output_nm}. \code{var_nm} should be a subset of the names returned by
\code{nc_list(base_dir)}.
}
