% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_layers.R
\name{nc_layers}
\alias{nc_layers}
\title{Open a subset of times from one or more NetCDF files}
\usage{
nc_layers(p, times = NULL, preload = TRUE, na_rm = FALSE)
}
\arguments{
\item{p}{character vector path to the nc file(s)}

\item{times}{vector of unique POSIXct times to match in the file(s)}

\item{preload}{logical indicating to load values into RAM}

\item{na_rm}{logical indicating to return only non-NA layers}
}
\value{
a SpatRaster with \code{length(t)} layers
}
\description{
Returns a SpatRaster with time-indexed layers in chronological order,
loaded from one or several NetCDF files. This wraps \code{terra::rast}, setting
the \code{lyrs} argument by matching \code{times} to times listed in the JSON file(s)
in subdirectory "/time".  If \code{times} is \code{NULL}, the function returns all
times found in the first file \code{p[1]}.
}
\details{
\code{p} can be a vector of paths, indicating to match \code{times} to the times in each
file. If a time is found in multiple files, the function returns the matching
layer from the first file (WRT to the order in \code{p}), and ignores the rest.
The returned SpatRaster will include one layer for all \code{times} that could be
matched, with different layers possibly coming from different files.

This functionality allows you to split time series into temporal chunks,
and store them in multiple files. Note that the function assumes all files
in \code{p} contain data for the same variable (variable names are ignored).

The function creates the JSON file(s) as needed, and assumes that
all of \code{times} can be found in the .nc file(s) at \code{p}. Available times
for the file at \code{p} can be listed with \code{terra::rast(p) |> terra::time()},
and this should match the "time" entries of \code{time_wx(p)}, assuming the JSONs
are up to date.

When \code{preload=TRUE} the function forces \code{terra} to copy all data values
into memory before returning the SpatRaster handle. Normally \code{terra} loads
values only when they are needed, but I have found that this can create
problems with file operations when there are a large number of times
(layers) in the file.

When \code{na_rm=TRUE}, only layers with all grid points observed (non-NA) are
searched. In this case, times that only match NA layers are reported
as "unmatched", and none of the NA layers are returned. When \code{na_rm=FALSE}
the first matching layer in \code{p} is returned for each time, and this may be
an NA layer (depending on the order of \code{p} and the contents of its files).
}
