% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_list.R
\name{nc_list}
\alias{nc_list}
\title{List the expected NetCDF files in the archive}
\usage{
nc_list(
  base_dir,
  rap = .nm_rap_export,
  gfs = .nm_gfs_export,
  var_rap = .var_rap_export,
  var_gfs = .var_gfs_export
)
}
\arguments{
\item{base_dir}{path to parent directory of GRIB storage subfolders}

\item{rap}{character vector of sub-directory names for RAC/RUC files}

\item{gfs}{character vector of sub-directory names for GFS files}

\item{var_rap}{list of character vectors, naming the RAP/RUC variables to include}

\item{var_gfs}{list of character vectors, naming the GFS variables to include}
}
\value{
character vector or list of paths (depends if \code{rap} and/or \code{gfs} are NULL)
}
\description{
This returns a list of paths to the NetCDF data files created by this package
with default settings. If either of \code{rap} or \code{gfs} is NULL, files from that
model are omitted.
}
\details{
Variable names are passed in a list, and can split into groups of files (not
necessarily having the same name) that are considered equivalent when loading
data. eg we consider 'pcp' and 'pcp_total' to be equivalent by including
\code{c('pcp', 'pcp_total')} in the first element of \code{var_rap} (see \code{?file_wx}).
}
