% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_resample.R
\name{nc_resample}
\alias{nc_resample}
\title{Create/update a NetCDF archive of resampled coarse resolution grids}
\usage{
nc_resample(
  var_nm,
  base_dir,
  input_nm = list(coarse = "coarse", fine = "fine"),
  output_nm = "coarse_resampled",
  r_fine = NULL,
  from = NULL,
  ...
)
}
\arguments{
\item{var_nm}{character vector or list of them, names of the variables to process}

\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{input_nm}{list of character vectors, namess of the source file subdirectories in \code{base_dir}}

\item{output_nm}{character name for the sub-directory to create.}

\item{...}{additional arguments passed to \code{terra::project}}

\item{SpatRaster}{optional target grid for resampling}
}
\value{
nothing, but possibly modifies the nc and JSON files in \code{file.path(base_dir, output_nm)}
}
\description{
This resamples the coarse resolution time series to get data at fine resolution,
writing results to NetCDF files in subdirectory \code{output_nm}. Only times found in
the coarse series but not the fine series are processed
}
\details{
Specify the coarse and fine sub-directories as named elements in list \code{input_nm}. Each
should be a character vector of sub-directories. These are checked for time series files
named \code{var_nm} (see \code{?nc_layers})

\code{var_nm} can be a list of character vectors, where each vector gives a set of variable
names to be treated as equivalent (eg 'pcp' and 'pcp_total'). The output variables are
assigned names from \code{names(var_nm)}.

The function creates sub-directory \code{output_nm} if it doesn't exist already. Existing
files in \code{output_nm} are updated by appending all times in \code{input_nm} occurring after
the last time in \code{output_nm}.

Coarse resolution grids are by default resampled by bilinear averaging (with GDAL).
Change this by setting named arguments in \code{...} to pass to \code{terra::project}.
}
