% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_update.R
\name{nc_update}
\alias{nc_update}
\title{Create or update a NetCDF version of a local RUC/RAP/GFS archive}
\usage{
nc_update(
  aoi,
  base_dir,
  output_nm = list(coarse = c("coarse"), fine = c("fine")),
  regex = .rap_regex,
  from = NULL,
  to = NULL,
  grib_dir = file_wx("grib", base_dir)
)
}
\arguments{
\item{aoi}{geometry object passed to \code{grib_idx} (area of interest)}

\item{base_dir}{character path to parent directory of \code{names(output_nm)}}

\item{output_nm}{list of character vectors, sub-directories in \code{base_dir} for the nc files}

\item{regex}{character vector passed to \code{grib_idx} (layer names)}

\item{from}{Date or vector of them, GRIB files for all earlier times are ignored}

\item{to}{Date or vector of them, GRIB files for all later times are ignored}

\item{grib_dir}{character path to GRIB directory (default is "base_dir/grib")}
}
\value{
nothing, but possibly writes to the nc and JSON files in \code{file.path(base_dir, output_nm)}
}
\description{
Most GRIB files fetched by \code{archive_update} are around 20-30 MB compressed,
so loading tens of thousands at once is extremely slow. To make the time series
easier to work with, this function extracts the subset that we need and saves it in
a format that can be loaded more quickly.
}
\details{
The data are merged into a single (multi-layer) NetCDF file for each of the variables
selected by \code{regex}. Only the sub-grid overlapping with \code{aoi} is copied.

Output files are given base names \code{names(regex)}, and extension '.nc', and
have layers named after the POSIXct time (as character, in GMT) at which the
forecast is valid. To see the paths of the output nc files, do:

\code{file_wx('nc', base_dir, output_nm, regex)}

RAP/RUC GRIBS come in two resolutions, so these are processed separately and saved
to separate sub-directories, named in \code{output_nm}. This means there are two output
nc files per variable: the primary 13km grid series ("fine"), and the much smaller
25km series ("coarse").

\code{output_nm} should be a list with entries 'coarse' and/or 'fine', each containing
one or more subdirectory names. This is to allow users to store a small nc file
for relatively new data separately from one or more "archived" nc files that are
unlikely to change very often. The function always writes to the first subdirectory
named in each of the elements of \code{output_nm}, but all subdirectories are checked
when determining if a time has been processed yet.

A date range for processing can be specified by passing either or both of \code{from}
and \code{to}. GRIB files outside of this date range are ignored. Different date ranges
can be specified for each variable, in which case \code{length(from)} and \code{length(to)}
should match \code{length(regex)}. By default, \code{from} and \code{to} are set to the the
earliest and latest GRIB available dates.
The script will be very slow on the initial run with many input GRIBs, but subsequent
calls to update an existing set of nc files (and JSONs) will be much faster, as
only the missing layers are read and copied, and the files to modify on disk are
relatively small.
}
