% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcp_update.R
\name{pcp_update}
\alias{pcp_update}
\title{Create/update a netCDF file with missing pcp_total times filled using pcp_large and pcp_small}
\usage{
pcp_update(
  base_dir,
  pcp_nm = "pcp",
  input_nm = list(coarse = "coarse", fine = "fine"),
  output_nm = c("coarse", "fine")
)
}
\arguments{
\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{pcp_nm}{a name for the output variable and file}

\item{input_nm}{list of character vectors, naming input subdirectories in \code{base_dir}}

\item{output_nm}{character vectors, naming an output subdirectory for each \code{input_nm} vector}
}
\value{
nothing, but possibly modifies the nc and JSON files in \code{file.path(base_dir, output_nm)}
}
\description{
This creates a copy of the "pcp_total" time series, named \code{pcp_nm}, containing only
times imputed using the sum of "pcp_large" (large scale precipitation) and "pcp_small"
(convective precipitation) whenever both are available, but "pcp_total" is not.
}
\details{
\code{input_nm} should be a named list character vector elements 'coarse' and/or 'fine', each
representing a batch of files to read a common variable from each resolution. Typical usage
reads several files for each variable, from different steps in the gap-filling workflow.

The function writes one output file for each of the vectors in \code{input_nm}, to the
location(s) in \code{output_nm}. The output file name is \code{paste0(pcp_nm, 'nc')}, so the
function will not allow you to set \code{pcp_nm} to any of the (in-use) names to avoid
confusion downstream (list them with \code{file_wx('nc', base_dir, output_nm)})

Layers with negatives are set to \code{NA} in the output file.

A JSON for the output is written to the "time" subdirectory. The function will update
existing files by only copying layers for times not already listed in the file.
}
