% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_fit.R
\name{run_spatial_fit}
\alias{run_spatial_fit}
\title{Fit a spatial model and return a list of results for use in \code{spatial_fit}}
\usage{
run_spatial_fit(p, X, n_max = 100, time_fit = NULL, pos = NULL)
}
\arguments{
\item{p}{character vector of input file paths to time series .nc files}

\item{X}{numeric matrix of spatial covariates (output of \code{space_X})}

\item{n_max}{integer > 0 number of layers to sample}

\item{time_fit}{character vector of POSIXct times to sample}

\item{pos}{logical indicating to exclude all-zero layers}
}
\value{
list of model fitting results
}
\description{
This is a helper function for \code{spatial_fit}. It fits a random subsample of
the observed data (from \code{p}) using the covariates in \code{X}. Generate \code{X} by passing
an example grid from \code{p} (along with a DEM) to \code{space_X}.
}
\details{
The function returns a list with the following
\itemize{
\item at : the UTC time of the function call
\item file : the files containing the sampled input data
\item pars : fitted parameters in a snapKrig-friendly list
\item bds : parameter bounds (lower, initial, fitted, upper)
\item betas : GLS estimates of the linear model coefficients
\item train : the times in the training set
}
}
