% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fit.R
\name{time_fit}
\alias{time_fit}
\title{Fit a temporal model to a NetCDF time series}
\usage{
time_fit(
  var_nm,
  base_dir,
  input_nm = "fine",
  daily_n = 5L,
  yearly_n = 5L,
  knots_n = 5L,
  p_max = 0.1,
  n_max = NA
)
}
\arguments{
\item{var_nm}{list of character vectors, the variable names to fit}

\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{input_nm}{character vector or list, subdirectories containing the .nc files to fit}

\item{daily_n}{integer number of Fourier term pairs for daily cycles}

\item{yearly_n}{integer number of Fourier term pairs for yearly cycles}

\item{knots_n}{integer number of knots to use for spline}

\item{p_max}{numeric the maximum proportion of missing points in the series}

\item{n_max}{integer maximum number of points to sample (NA to select all)}
}
\value{
returns nothing but modifies the files in "model" subdirectory of "input_nm"
}
\description{
Fit a temporal model to a NetCDF time series
}
