% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_impute.R
\name{time_impute}
\alias{time_impute}
\title{Fill temporal gaps using AR(2) model and seasonal predictor}
\usage{
time_impute(
  var_nm,
  base_dir,
  input_nm = "fine",
  output_nm = "completed",
  model_dir = base_dir,
  model_nm = .nm_model,
  from = NULL,
  to = NULL,
  quiet = FALSE
)
}
\arguments{
\item{var_nm}{character vector or list of them, the input variables to process}

\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{input_nm}{character vector or list, subdirectories containing the .nc files to process}

\item{output_nm}{character, sub-directory to write the output nc files}

\item{model_nm}{character, sub-directory to look for fitted model files}

\item{from}{POSIXct time, start of time range to process}

\item{to}{POSIXct time, end of time range to process (passed as "until" to \code{archive_pad})}

\item{quiet}{logical indicating to suppress console output}
}
\value{
nothing, but writes output to nc files in \code{output_nm}
}
\description{
Fill temporal gaps using AR(2) model and seasonal predictor
}
