% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_wx.R
\name{time_json}
\alias{time_json}
\title{Open JSON file(s) associated with NetCDF time series}
\usage{
time_json(nc_path)
}
\arguments{
\item{nc_path}{character vector, path(s) to the NetCDF file(s)}
}
\value{
a list of list(s) with vectors 'na' (integer), 'time', 'time_na', 'time_obs'
}
\description{
To improve access and write times for large NetCDF files, we use a JSON file
to index times and layers containing NAs. This JSON is located in the "time"
sub-directory (relative to the NetCDF file) and has the same name up to the
file extension.
}
\details{
This function returns the data from the JSON in a list, along with a pair of
derived time vectors (see \code{?time_wx}). If the JSON file is not found, the
function instead returns NA (see \code{?write_time_json}). If a vector of paths is
supplied, the function loops over them and returns all results in a list.

For consistency with the vectorized case, the function returns a length-1
list when a single path is supplied to \code{nc_path}.
}
