% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_wx.R
\name{time_nc}
\alias{time_nc}
\title{Read a NetCDF or GeoTIFF time series file to find times and layers with NAs}
\usage{
time_nc(r)
}
\arguments{
\item{r}{SpatRaster or character vector of paths to NetCDF or GeoTIFF file(s)}
}
\value{
a list with vectors 'na' (integer), 'time', 'time_na', 'time_obs'
}
\description{
For SpatRaster \code{r}, the function returns a list of four vectors:
}
\details{
\itemize{
\item "na": index of layers containing any number of NAs
\item "time" : all times as POSIXct strings, in same order as layers
\item "time_obs" : elements of "time" whose layers are not listed "na"
\item "time_na" : elements of "time" whose layers are listed "na"
}

or \code{NA}, if the file is not found. All times are assumed to be in the UTC time zone.

If \code{r} is a vector of paths, the function opens each one in a loop
and returns results as a list (same length as \code{r}) of lists with
the above structure. Top-level names correspond to the path
that was read.

This function can be slow with large files, so we recommend calling
once only, then caching results in a JSON (see \code{?write_time_json})
where they can be retrieved much more quickly using \code{time_json}.
}
