% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_step.R
\name{time_step}
\alias{time_step}
\title{Find the minimal time difference between subsequent observations in a time series}
\usage{
time_step(grib_df, times = "posix_pred", quiet = FALSE, more = FALSE)
}
\arguments{
\item{grib_df}{data frame with POSIXct times in column 'times'}

\item{times}{character column name of times}

\item{quiet}{logical, if TRUE the function warns about NAs}

\item{more}{logical, if TRUE the function returns a data frame}
}
\value{
numeric (hours) or copy of \code{grib_df} with NA times removed and two columns appended
}
\description{
Returns the step size in hours or, if \code{more=TRUE}, appends the columns "ts_hour" (the
number of hours since the first observation) and "interval" (the difference in
"ts_hour" between a given observation and the previous). The step size is the least
nonzero "interval".
}
\details{
NA times are removed from the results. The default \code{quiet=FALSE} warns if this happens.
}
