% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_X.R
\name{time_X}
\alias{time_X}
\title{Return a data matrix of seasonal linear predictors for a time series}
\usage{
time_X(
  t_obs,
  intercept = TRUE,
  na_rm = FALSE,
  knots_t = NULL,
  knots_n = 5L,
  daily_n = 5L,
  yearly_n = 5L
)
}
\arguments{
\item{t_obs}{POSIXct vector, the observed times in the time series}

\item{intercept}{logical indicating to include an intercept column (of \code{1}s)}

\item{na_rm}{logical, omits rows from the output where the corresponding time is unobserved}

\item{daily_n}{integer, the number of fourier term pairs to use for daily cycles}

\item{yearly_n}{integer, the number of fourier term pairs to use for yearly cycles}

\item{freq_year}{integer, the number of observations per year (365.25 days)}

\item{default_n}{integer, copied to \code{daily_n} and/or \code{yearly_n} when they are NA}
}
\value{
a matrix of covariates
}
\description{
This creates a data matrix for daily and yearly cycles in a regular time series
observed at times \code{t_obs}. With \code{na_rm=TRUE}, the output has a row for each
\code{t_obs}. Otherwise it has a row for each time in the sequence of frequency
\code{freq_year} spanning \code{t_obs}.
}
