% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_extract.R
\name{which_lyr}
\alias{which_lyr}
\title{Find the index of first match to each element in pattern against a set of strings}
\usage{
which_lyr(available, pattern = NULL, na.rm = FALSE, quiet = FALSE)
}
\arguments{
\item{available}{character vector of strings to match against}

\item{pattern}{character vector of regular expressions}

\item{na.rm}{logical, if TRUE omit unmatched results}

\item{quiet}{logical, if TRUE the function won't warn about multiple matches}
}
\value{
integer vector
}
\description{
Helper for \code{grib_idx} and \code{grib_extract}, where \code{available} are the layer names in
a GRIB file and we want to select exactly one of them for each pattern.
}
\details{
If pattern is NULL the function instead returns all indices 1...length(\code{available})
}
