% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnd_update.R
\name{wnd_update}
\alias{wnd_update}
\title{Create/update a netCDF file with wind speed, computed from its components}
\usage{
wnd_update(
  base_dir,
  wnd_nm = .var_wnd,
  uv_nm = .var_wnd_uv,
  input_nm = .nm_complete_rap,
  output_nm = wnd_nm
)
}
\arguments{
\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{wnd_nm}{a name for the output variable and file}

\item{input_nm}{list of character vectors, naming input subdirectories in \code{base_dir}}

\item{output_nm}{character vectors, naming an output subdirectory for each \code{input_nm} vector}
}
\value{
nothing, but possibly modifies the nc and JSON files in \code{file.path(base_dir, output_nm)}
}
\description{
This creates a new variable \code{wnd_nm} reporting the norm (speed) of the
wind vector, computed from the two components (u and v). These components should
be found in the length-2 names vector \code{uv_nm}. Specify the directory names in
which to look for these variables with \code{input_nm} (see also \code{?pcp_update}).
}
\details{
The function writes a single NetCDF file to the directory named in \code{output_nm}.
This file is named \code{paste0(wnd_nm, '.nc')} like the other variables, so the
function will not allow you to set \code{wnd_nm} to any of the (in-use) variable
names (list them with \code{file_wx('nc', base_dir, output_nm)})

An output time (layer) is generated for each time in which both wind component
variables are observed in at least one the files in \code{input_nm}. The function only
appends those times not already found in the output file when overwriting it.
A JSON file in the "time" subdirectory is also created/updated (see \code{?nc_write}).
}
