% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_downscale}
\alias{workflow_downscale}
\title{Construct down-scaled (fine resolution) estimates of daily spatial grid data}
\usage{
workflow_downscale(
  project_dir,
  down = 100,
  dem_path = NULL,
  poly_path = NULL,
  from = NULL,
  to = NULL
)
}
\description{
Wrapper for \code{nc_downscale} to produced down-scaled versions of the daily outputs
produced by \code{workflow_daily}.
}
\details{
This requires a digital elevation model (DEM) with data in meters. This is expected
in the GeoTIFF file "elev_m.tif" in \code{project_dir}, but you can specify an alternate
name or location with \code{dem_path} (passed to \code{terra::rast}).

The DEM is warped to the target coordinate reference system by bilinear averaging
Output is written to the sub-directory \code{.nm_export} of \code{project_dir}. By default the
function overwrites layers starting from 10 days before the latest date in the existing
output files (see \code{?workflow_daily}). Change this by setting start/end dates
in \code{from}/\code{to}.
}
