% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_export}
\alias{workflow_export}
\title{Exported completed time series to daily aggregate values and write to disk}
\usage{
workflow_export(project_dir, write_csv = TRUE)
}
\arguments{
\item{project_dir}{character path to the project root directory}
}
\value{
returns nothing but possible writes to \code{project_dir}
}
\description{
Exported completed time series to daily aggregate values and write to disk
}
