% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_extract}
\alias{workflow_extract}
\title{Aggregate data over polygons and write results to CSV}
\usage{
workflow_extract(
  project_dir,
  poly_path = NULL,
  fun = NULL,
  from = NULL,
  to = NULL
)
}
\arguments{
\item{project_dir}{character path to the project root directory}

\item{poly_path}{character path to polygons to summarize}

\item{fun}{character, function name (probably either mean, min, or max)}

\item{from}{POSIXct start of time range}

\item{to}{POSIXct end of time range}
}
\value{
returns nothing but possible writes to sub-directory \code{.nm_export}
}
\description{
This processes the output from workflow_downscale to extract aggregate
(summary) values over whole areas, defined by the polygons in \code{poly_path}.
}
\details{
Each time step of the input appears as a row in the output. The function 'fun'
is applied to all values within each polygon to form the columns, and these are
numbered in the same order as the polygons in \code{poly_path}.
}
