% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_impute_rap}
\alias{workflow_impute_rap}
\title{Impute missing time layers}
\usage{
workflow_impute_rap(project_dir)
}
\arguments{
\item{project_dir}{character path to the project root directory}
}
\value{
returns nothing but possible writes to \code{project_dir}
}
\description{
This creates the \code{.nm_complete} sub-directory of \code{project_dir} and
fills it with NetCDF files containing imputed layers for times missing
from the time series named in \code{.nm_resample_rap}.
}
\details{
Users must first run \code{workflow_update_rap} and \code{workflow_fit_temporal} at
least once before calling this function.

After running \code{workflow_update_rap} and \code{workflow_impute_rap}, users should
be able to access an up-to-date completed RAP time series (with no NA layers) by
passing \code{.nm_complete_rap} to \code{file_wx} then \code{nc_layers}.
}
