% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_list}
\alias{workflow_list}
\title{List all times and variables available locally in the RAP/RUC/GFS archive}
\usage{
workflow_list(project_dir, quiet = FALSE)
}
\arguments{
\item{project_dir}{character path to the project root directory}

\item{quiet}{logical suppresses console messages}
}
\value{
a list of character vectors, paths to individual NetCDF files
}
\description{
This checks the sub-directories of \code{project_dir} (both 'rap' and 'gfs') for
NetCDF data on the variables named in \code{.nm_output_var}, and reports some stats
to the console
}
