% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_update_gfs}
\alias{workflow_update_gfs}
\title{Update the GFS archive}
\usage{
workflow_update_gfs(project_dir, n_ahead = 3)
}
\arguments{
\item{project_dir}{character path to the project root directory}

\item{n_ahead}{integer, predictions run ahead of GFS data by this many days}
}
\value{
returns nothing but possible writes to \code{project_dir}
}
\description{
Download the latest files, convert the times needed to NetCDF. This is a wrapper
for a sequence of calls to \code{archive_update}, \code{nc_update} in that order.
}
\details{
Users should first run \code{workflow_update_rap} and \code{workflow_impute_rap}, in that
order, to update the RAP archive. The function checks the output from those functions
(sub-directories named in \code{.nm_complete_rap}) and identifies, for each variable. the
latest time available. Only times occurring after this point are extracted from GFS.

A geometry object must be supplied to define the bounding box to fetch from GFS.
The function expects to find it in the file 'aoi.geojson' in \code{project_dir}.

This will create/modify the 'gfs' sub-directory of \code{project_dir}, writing NetCDF
versions of the variables named in \code{.gfs_regex}.
}
