% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_update_rap}
\alias{workflow_update_rap}
\title{Update the RUC/RAP archive}
\usage{
workflow_update_rap(project_dir)
}
\arguments{
\item{project_dir}{character path to the project root directory}
}
\value{
Returns nothing but possible writes to \code{project_dir}
}
\description{
Download the latest files, convert to NetCDF and replace some of the
missing layers. This is a wrapper for a sequence of calls to \code{archive_update},
\code{nc_update}, \code{pcp_update}, then \code{nc_resample}, in that order.
}
\details{
A geometry object must be supplied to define the bounding box for NetCDF data.
This function expects to find it in the file 'aoi.geojson' in \code{project_dir}.

This creates/modifies several sub-directories of \code{project_dir} (named in
\code{.nm_resample_rap}), writing NetCDF versions of the variables named in
\code{.rap_regex} and \code{.var_pcp}.
}
