% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_wnd_rap}
\alias{workflow_wnd_rap}
\title{Create wind speed layer}
\usage{
workflow_wnd_rap(project_dir)
}
\arguments{
\item{project_dir}{character path to the project root directory}
}
\value{
returns nothing but possible writes to \code{project_dir}
}
\description{
This creates the \code{wnd} sub-directory of \code{project_dir} and writes a NetCDF file
for variable "wnd" after computing it from components "wnd_u" and "wnd_v".
}
\details{
Users should either call this function after running the workflow up to
(and including) \code{workflow_impute_rap} to get a complete wind speed time series
from the gap-filled components.
}
