% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_write.R
\name{write_time_json}
\alias{write_time_json}
\title{Create or update the JSON file associated with a NetCDF time series}
\usage{
write_time_json(nc_path, r = NULL, append = TRUE)
}
\arguments{
\item{nc_path}{character path to the NetCDF file}

\item{r}{SpatRaster or NULL, containing times to add}

\item{append}{logical indicating to append times rather than overwrite}
}
\value{
logical indicating if the file was modified
}
\description{
This manages the JSON file read by \code{time_json}. To get the path to this
file, do \code{nc_path |> time_json() |> names()}. This function will create both
the file and its parent directory (if they don't exist already). It only modifies
the JSON and never the NetCDF file.
}
\details{
Note that the JSON contains only the fields "na" and "time" and not the derived
fields "time_na" and "time_obs" (see \code{?time_wx}).

Call the function with default arguments to initialize the JSON for an existing
NetCDF file. This involves loading all data into memory, so it can be very slow when
there are numerous time points.

If \code{r} is supplied along with \code{append=TRUE}, the function creates/updates the JSON
to make it consistent with the result of concatenating the layers in \code{nc_path} with
the layers in \code{r}. Only times occurring after the last time in \code{nc_path} are copied
from \code{r}

Set \code{append=FALSE} to delete an existing JSON file and replace it with all of the
layer info for SpatRaster \code{r},

If \code{nc_path} doesn't exist, the function returns \code{FALSE} and writes nothing
}
