% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_downscale.R
\name{.nc_spatial_aggregate}
\alias{.nc_spatial_aggregate}
\title{Aggregate raster data over a set of polygons}
\usage{
.nc_spatial_aggregate(r, fun = "mean", poly_list = NULL)
}
\arguments{
\item{r}{SpatRaster with extent covering \code{poly_list}}

\item{fun}{character name of the function to use for aggregation (or NULL for no aggregation)}

\item{poly_list}{list of sf POLYGON objects to use for cropping output (or NULL for no cropping)}
}
\value{
either a SpatVector, a list of them
}
\description{
This aggregates the values in SpatRaster \code{r} by masking to the polygon(s) in \code{poly_list}
and applying the function \code{fun} to each subset.
}
\details{
\code{r} can have multiple layers. A single numeric is returned for each polygon, forming
a vector of aggregate values at different times. When \code{poly_list} has length 1 this is
returned as a vector, otherwise it is returned as a matrix with one row for each polygon.

\code{fun} specifies the name of the aggregation function to use - typical usage will set one
of 'mean' 'max' and 'min' (see \code{?terra::extract}). Alternatively, set \code{fun=NULL} to return
a list of SpatRasters cropped to \code{poly_list} (one per polygon). When \code{fun} is not \code{NULL},
this list is passed (along with \code{fun}) to \code{terra::extract} and the result is reshaped as
described above.
}
