% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_export.R
\name{nc_aggregate}
\alias{nc_aggregate}
\title{Aggregate sub-daily data to daily}
\usage{
nc_aggregate(p, fun = "mean", tz = "UTC", origin_hour = 0L)
}
\arguments{
\item{p}{character vector path to the nc file(s)}

\item{fun}{function, one of "mean", "min", or "max"}

\item{tz}{character time zone for \code{origin_hour}}

\item{origin_hour}{integer, steps are aligned to start at this hour of the day}
}
\value{
SpatRaster, the aggregated data
}
\description{
Returns a list of SpatRasters, one for each day in the time series
\code{p}. Times within a given day are aggregated using the function
\code{fun} should be a function (not a string naming it), such as \code{mean}
or \code{max} (see \code{?terra::app}).
}
\details{
This has not been tested on incomplete data
}
