% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_aggregate.R
\name{nc_aggregate_time}
\alias{nc_aggregate_time}
\title{Aggregate the completed 2-hourly time series to daily}
\usage{
nc_aggregate_time(
  base_dir,
  var_nm = NULL,
  output_nm = .nm_daily,
  fun = "mean",
  tz = "UTC",
  origin_hour = 0L
)
}
\arguments{
\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{var_nm}{character vector, the variable(s) to export (NULL to export all)}

\item{output_nm}{character, the sub-directory name for output}

\item{fun}{function, a function name like "mean", "min", or "max" (see \code{?.nc_aggregate_time})}

\item{tz}{character time zone for \code{origin_hour}}

\item{origin_hour}{integer, steps are aligned to start at this hour of the day}
}
\value{
vector of file paths written to the \code{output_nm} directory
}
\description{
This creates a (single) NetCDF file in \code{output_nm} for each of the variables named in
\code{var_nm} by applying the function named in \code{fun} to aggregate by day. \code{var_nm} should
be a subset of the names returned by \code{nc_list(base_dir)}.
}
\details{
If \code{fun=NULL}, the function does no aggregation and writes all time points.

If \code{fun} is not \code{NULL} then the last three arguments are passed to \code{.nc_aggregate_time} to control
the alignment of the aggregation window. \code{fun} specifies the function to use for combining
times within the window (see \code{?.nc_aggregate_time}).  Set \code{tz} to the desired output time
zone, and leave \code{origin_hour=0} to have each day begin at 12AM (in time zone \code{tz}).

File names for aggregate data are given the suffix \verb{_daily_<fun>} - eg with \code{fun='mean'},
the output "tmp.nc" becomes "tmp_daily_mean.nc".
}
