% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_chunk.R
\name{nc_chunk}
\alias{nc_chunk}
\title{Expand a directory path to get a set of paths to the NetCDF files within}
\usage{
nc_chunk(p, year = NULL)
}
\arguments{
\item{p}{character path to a file or directory with extension '.nc'}

\item{year}{(optional) character or integer vector of (4-digit) years}
}
\value{
character vector of file path(s) to NetCDF file(s) associated with \code{p}
}
\description{
It can be convenient to split a long time series into year-long chunks
and write each chunk to its own NetCDF (.nc) file. For variable "var", these
files are written to a directory named "var.nc". This function looks for files
in this directory and returns their paths.
}
\details{
With the default \code{year=NULL} the function scans for files at \code{p} and returns what
it finds there: either \code{p} itself if the path points to a NetCDF file, or a vector
of paths to all NetCDF files inside the directory \code{p}. If the directory exists but is
empty, the function returns \code{character(0)}.

If \code{year} is supplied, the function returns file paths for each of \code{unique(year)},
without checking if the parent directory exists.

The expected file name for a chunk is "var_year.nc", with "year" the 4-digit year
of all times in the file. For example "tmp.nc/tmp_2005.nc" holds observations of
temperature from 2005.
}
