% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_downscale.R
\name{nc_downscale}
\alias{nc_downscale}
\title{Down-scale a spatio-temporal time series in a set of NetCDF files}
\usage{
nc_downscale(
  base_dir,
  dem,
  down = 100,
  input_nm = .nm_daily,
  model_nm = .nm_model,
  output_nm = .nm_down,
  var_nm = .var_daily,
  poly_out = NULL,
  edge_buffer = NULL,
  from = NULL,
  to = NULL,
  file_ext = "both"
)
}
\arguments{
\item{base_dir}{path to parent directory of GRIB storage subfolder}

\item{dem}{SpatRaster of elevation data at finer resolution than the input data}

\item{down}{integer, the down-scaling factor}

\item{input_nm}{character, the sub-directory name of input to process}

\item{model_nm}{character, the sub-directory name where the model files can be found}

\item{output_nm}{character, the sub-directory name for output}

\item{var_nm}{character vector, the name(s) of the variable to process}

\item{poly_out}{sfc object, polygon of interest}

\item{edge_buffer}{numeric, length in metres to buffer input data (see details)}

\item{from}{Date, the first date of the sequence to process}

\item{to}{Date, the last date of the sequence to process}

\item{file_ext}{character, either 'tif', 'nc', or 'both'}
}
\value{
returns nothing but possibly modifies the NetCDF data in \code{output_nm}
}
\description{
This is a wrapper for \code{.nc_downscale} that down-scales - ie predicts at finer
resolution - the gridded time series data in \code{input_nm}. By default it writes
a set of yearly GeoTIFF files in the sub-directory \code{output_nm} of \code{base_dir},
with one set for each of the variables named in \code{var_nm}.
}
\details{
Down-scaling happens by universal kriging using the fitted spatial covariance
parameters saved in the folder named in \code{model_nm} (generate this using
\code{space_fit}), and using covariates generated from the digital elevation model
in \code{dem} with values in metres.

Argument \code{down} is the factor by which to reduce grid line spacing. This should
be an integer greater than 1. eg. \code{down=2} "doubles the resolution" by inserting
a new grid line in between each existing pair of adjacent ones. This produces an
output with approximately 4X (ie \code{down^2}) as many grid points as the input.

The function expects \code{down} to be used consistently within a given output
subdirectory. If you want to change the output resolution, you will need to delete
any existing output or else specify a new \code{output_nm}.

Use \code{from} and \code{to} to specify a date range to update (inclusive), or leave them \code{NULL}
to use a default range. The default range is meant to cover layers originating from GFS
(allowing them to be can be replaced by newly added RAP layers, or more recently released
GFS forecasts).

The default for \code{to} is always the latest available date in the input. The default for
\code{from} is 10 days before the latest date found in the existing output files. If there
are no existing outputs, the default is set to the earliest available date in the input.

Define an area of interest with \code{poly_out}. The function conditions its
predictions on a subset of the input grid covering \code{poly_out}, buffered by
the distance \code{edge_buffer} (in m). By default this is set to the diagonal length
of a single grid cell. Set this to a larger number, like the side length of your
domain, to ensure that all observations are considered in predictions, reducing
edge effects.
}
