% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_aggregate.R
\name{nc_export}
\alias{nc_export}
\title{Export completed time series, optionally aggregating to daily}
\usage{
nc_export(base_dir, p_all = NULL, output_nm = fun, agg = NULL)
}
\arguments{
\item{agg}{list with entries "fun", "tz", "origin_hour" to pass to \code{nc_aggregate}}

\item{var_nm}{character vector, the variable(s) to export}
}
\value{
returns nothing but writes to the \code{output_nm} directory
}
\description{
Export completed time series, optionally aggregating to daily
}
