% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_write.R
\name{nc_write}
\alias{nc_write}
\title{Create or modify a NetCDF file by appending new time layers from SpatRaster}
\usage{
nc_write(r, p)
}
\arguments{
\item{r}{SpatRaster with POSIXct vector \code{terra::time(r)}, the data to write}

\item{p}{character path the (.nc) time series data file/directory to write}
}
\value{
vector of POSIXct times, the layers added to the file
}
\description{
This creates or updates an existing NetCDF file with new times (layers) in the
supplied SpatRaster \code{r}.
}
\details{
\code{r} must have a time for each layer (check \code{terra::time(r)}). Only new times
are copied - ie those occurring after the latest time in the file at \code{p}. To
add/modify earlier times you will need to delete the file and start over.

The function assumes that \code{p} is a valid path to a NetCDF file (.nc), and that
the raster grid in \code{r} is compatible with the file at \code{p} in terms of dimensions,
projection, etc.

When overwriting an existing file, the function initially writes to a temporary
file in the same directory. At the end of the function call the old file is
deleted and the temporary one is renamed to take its place. If something goes
wrong and the function halts, this temporary file can be safely deleted.

After writing to the .nc file, the function creates/updates the JSON attributes
file in the 'time' subdirectory. If something goes wrong with this file, it can
be safely deleted then rebuilt with \code{time_wx(p)}.
}
