% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_wx.R
\name{time_wx}
\alias{time_wx}
\title{Find index of NA layers and times available in a (set of) NetCDF files(s)}
\usage{
time_wx(nc_path, join = TRUE, collapse = TRUE)
}
\arguments{
\item{nc_path}{character vector, path(s) to the NetCDF file(s)}

\item{join}{logical, when TRUE the results from all \code{nc_path} are combined}

\item{collapse}{logical, if TRUE and output list has length 1, it is unlisted}
}
\value{
a list with vectors 'na' (integer), 'time', 'time_na', 'time_obs' (or a list of them)
}
\description{
The function returns a list of vectors indicating the times (layers) available
in one or more NetCDF files (at \code{nc_path}), along with information about NAs.
}
\details{
If \code{nc_path} is a single path, or if \code{join=TRUE}, the function returns in a list:
\itemize{
\item "na": index of layers containing any number of NAs
\item "time" : all times as POSIXct strings, in same order as layers
\item "time_obs" : elements of "time" whose layers are not listed "na"
\item "time_na" : elements of "time" whose layers are listed "na"
}

If \code{nc_path} is a vector of multiple paths and \code{join=FALSE}, the function returns
a list of results of with the above structure (one per path).

\code{join=TRUE} allows a time series to be split into different files, each with a
(possibly) different chunk of times. The function checks all paths and joins the
times that it finds, returning all results in a single list. If a time is observed
in any of the paths in \code{nc_path}, it is marked as observed in the output.

If a NetCDF file is listed in \code{nc_path} but not not found on disk, it is dropped
from the results. When looking up times, the function calls \code{time_json} first, then
the much slower \code{time_nc} only if the JSON is not found. In that case the function
calls \code{write_time_json} to create it.

If none of the paths in \code{nc_path} point to an existing file, the function returns
an empty list
}
