% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_daily}
\alias{workflow_daily}
\title{Exported completed time series to daily aggregate values and write to disk}
\usage{
workflow_daily(project_dir, from = NULL, to = NULL, write_csv = FALSE)
}
\arguments{
\item{project_dir}{character path to the project root directory}

\item{from}{Date start of date range (default is earliest available)}

\item{to}{Date end of date range (default is latest available)}

\item{write_csv}{logical, if TRUE a CSV file and geoJSON are written in addition to the NetCDF}
}
\value{
returns nothing but possible writes to \code{project_dir}
}
\description{
Wrapper for \code{nc_aggregate}
}
\details{
This prepares five output variables aggregated to daily average, or maximum or minimum.
The pairing of variable names and aggregation functions is set up in the global
constant \code{.var_daily_pairs}
}
