% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_fit_spatial}
\alias{workflow_fit_spatial}
\title{Fit a spatial model to daily aggregate time series, or load it from a file}
\usage{
workflow_fit_spatial(
  project_dir,
  dem_path = NULL,
  from_file = "spatial_model.zip"
)
}
\arguments{
\item{project_dir}{character path to the project root directory}

\item{from_file}{character file name of zip containing previously fitted model}
}
\value{
returns nothing but possible writes to \code{project_dir}
}
\description{
A wrapper for \code{space_fit}. This reads its input from the sub-directory
named \code{.nm_daily} and writes its output to \code{.nm_spatial_model} (see \code{?space_fit}).
}
\details{
Fitting requires a digital elevation model (DEM) with data in meters. This is expected
in the GeoTIFF file "elev_m.tif" in \code{project_dir}, but you can specify an alternate
name or location with \code{dem_path} (passed to \code{terra::rast}).

Alternatively, load a previously fitted model from a zip archive by providing
the file name in \code{from_file}. This zip should contain the contents of
\code{.nm_spatial_model} generated from a \code{space_fit} call on a similar dataset

Call this function at least once after running \code{daily} for the first time if
you plan on running \code{workflow_downscale} and/or \code{workflow_extract}
}
