% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_fit_temporal}
\alias{workflow_fit_temporal}
\title{Fit a temporal model to the RUC/RAP times series}
\usage{
workflow_fit_temporal(project_dir)
}
\arguments{
\item{project_dir}{character path to the project root directory}
}
\value{
Returns nothing but possible writes to \code{project_dir}
}
\description{
A wrapper for \code{time_fit}. This reads its input from the sub-directories
named in \code{.nm_resample_rap} and writes its output to the first of these
sub-directories (for each variable). See \code{?time_fit}
}
\details{
Call this function at least once after running \code{workflow_update_rap} for
the first time (and before running \code{workflow_impute_rap})
}
