% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_write.R
\name{write_time_json}
\alias{write_time_json}
\title{Create or update the JSON file associated with a NetCDF time series}
\usage{
write_time_json(nc_path, r = NULL)
}
\arguments{
\item{nc_path}{character path to a single NetCDF file}

\item{r}{SpatRaster or NULL, containing times to add}
}
\value{
logical indicating if the file was modified
}
\description{
This manages the JSON file read by \code{time_json}. To get the path to this
file, do \code{nc_path |> time_json() |> names()}. This function will create both
the file and its parent directory (if they don't exist already). It only modifies
the JSON and never the NetCDF file.
}
\details{
Note that the JSON contains only the fields "na" and "time" and not the derived
fields "time_na" and "time_obs" returned by \code{time_nc} and \code{time_wx}.

If \code{r} is supplied, the function indexes the SpatRaster. If \code{r} is \code{NULL}, the
function instead indexes all layers in the file at \code{nc_path}. This can be very slow
if there are many time points.

If \code{nc_path} doesn't exist, the function returns \code{FALSE} and writes nothing
}
